golang assignment1 refactor
