+++ pankaj/golang/assignment1/go.mod
@@ -1,0 +1,3 @@
+module github.com/pankaj9310/go-k8s-training/pankaj/golang/assignment1
+
+go 1.15
